requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
astroid==1.5.2            # via pylint
colorama==0.3.9           # via libkeepass
easypysmb==1.4.2          # via pykeepass
isort==4.2.5              # via pylint
lazy-object-proxy==1.3.1  # via astroid
libkeepass==0.2.0         # via pykeepass
lxml==3.7.3               # via libkeepass
mccabe==0.6.1             # via pylint
pyasn1==0.2.3             # via pysmb
pycrypto==2.6.1           # via libkeepass
pykeepass==2.6.1
pylint==1.7.1
pypdf2===1.26.0
pysmb==1.1.19             # via easypysmb
pyyaml==3.12
selenium==3.4.1
six==1.10.0               # via astroid, pylint
wrapt==1.10.10            # via astroid
